+++ broker-core/src/main/java/io/zeebe/broker/exporter/debug/DebugHttpExporter.java
@@ -18,9 +18,9 @@
 package io.zeebe.broker.exporter.debug;
 
 import io.zeebe.broker.system.configuration.ExporterCfg;
+import io.zeebe.exporter.api.Exporter;
 import io.zeebe.exporter.api.context.Context;
 import io.zeebe.exporter.api.record.Record;
-import io.zeebe.exporter.api.spi.Exporter;
 import org.slf4j.Logger;
 
 public class DebugHttpExporter implements Exporter {
