+++ broker-core/src/test/java/io/zeebe/broker/exporter/repo/ExporterRepositoryTest.java
@@ -25,9 +25,9 @@ import io.zeebe.broker.exporter.util.ControlledTestExporter;
 import io.zeebe.broker.exporter.util.JarCreatorRule;
 import io.zeebe.broker.exporter.util.TestJarExporter;
 import io.zeebe.broker.system.configuration.ExporterCfg;
+import io.zeebe.exporter.api.Exporter;
 import io.zeebe.exporter.api.context.Context;
 import io.zeebe.exporter.api.record.Record;
-import io.zeebe.exporter.api.spi.Exporter;
 import java.io.File;
 import java.io.IOException;
 import java.util.HashMap;
