+++ broker-core/src/test/java/io/zeebe/broker/exporter/util/ControlledTestExporter.java
@@ -17,10 +17,10 @@
  */
 package io.zeebe.broker.exporter.util;
 
+import io.zeebe.exporter.api.Exporter;
 import io.zeebe.exporter.api.context.Context;
 import io.zeebe.exporter.api.context.Controller;
 import io.zeebe.exporter.api.record.Record;
-import io.zeebe.exporter.api.spi.Exporter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Consumer;
