+++ qa/integration-tests/src/test/java/io/zeebe/broker/it/DataDeleteTest.java
@@ -19,9 +19,9 @@ import io.zeebe.broker.system.configuration.BrokerCfg;
 import io.zeebe.broker.system.configuration.DataCfg;
 import io.zeebe.broker.system.configuration.ExporterCfg;
 import io.zeebe.broker.test.EmbeddedBrokerRule;
+import io.zeebe.exporter.api.Exporter;
 import io.zeebe.exporter.api.context.Controller;
 import io.zeebe.exporter.api.record.Record;
-import io.zeebe.exporter.api.spi.Exporter;
 import io.zeebe.protocol.intent.MessageIntent;
 import io.zeebe.test.util.TestUtil;
 import java.io.File;
