+++ test/src/main/java/io/zeebe/test/exporter/ExporterIntegrationRule.java
@@ -25,9 +25,9 @@ import io.zeebe.broker.system.configuration.ExporterCfg;
 import io.zeebe.client.ClientProperties;
 import io.zeebe.client.api.subscription.JobHandler;
 import io.zeebe.client.api.subscription.JobWorker;
+import io.zeebe.exporter.api.Exporter;
 import io.zeebe.exporter.api.record.Record;
 import io.zeebe.exporter.api.record.value.IncidentRecordValue;
-import io.zeebe.exporter.api.spi.Exporter;
 import io.zeebe.model.bpmn.Bpmn;
 import io.zeebe.model.bpmn.BpmnModelInstance;
 import io.zeebe.protocol.intent.IncidentIntent;
