+++ test/src/main/java/io/zeebe/test/exporter/ExporterTestHarness.java
@@ -18,10 +18,10 @@ package io.zeebe.test.exporter;
 import com.moandjiezana.toml.Toml;
 import io.zeebe.broker.system.configuration.BrokerCfg;
 import io.zeebe.broker.system.configuration.ExporterCfg;
+import io.zeebe.exporter.api.Exporter;
 import io.zeebe.exporter.api.context.Configuration;
 import io.zeebe.exporter.api.context.Controller;
 import io.zeebe.exporter.api.record.Record;
-import io.zeebe.exporter.api.spi.Exporter;
 import io.zeebe.test.exporter.record.MockRecord;
 import io.zeebe.test.exporter.record.MockRecordMetadata;
 import io.zeebe.test.exporter.record.MockRecordStream;
